Pulled out getString() to file
