+++ apps/www/pages/api/suggestions/index.ts
@@ -1,8 +1,9 @@
 import { normalizeQuery, QueryMatches } from "@phylopic/api-models"
-import { getSortIndex } from "@phylopic/ui" // :TODO: move to utils
+import { getSortIndex } from "@phylopic/ui"; // :TODO: move to utils
 import { createSearch, stringifyNormalized } from "@phylopic/utils"
 import axios from "axios"
 import { NextApiHandler } from "next"
+import getString from "~/routes/getString"
 const index: NextApiHandler = async (req, res) => {
     try {
         const prefix = getString(req.query.q)
@@ -29,15 +30,6 @@ type Suggestion = Readonly<{
     term: string
     url: string
 }>
-const getString = (value: string | string[] | undefined) => {
-    if (typeof value === "string") {
-        return value
-    }
-    if (value === undefined) {
-        return ""
-    }
-    return value[0]
-}
 const getSuggestions = async (prefix: string): Promise<readonly Suggestion[]> => {
     const suggestionBatches: ReadonlyArray<readonly Suggestion[]> = await Promise.all([
         getPhyloPicSuggestions(prefix),
